einkommensteuer

+++ TPE/uebungen/src/tpe/fruh_razzaq_jando/pue2/Einwohner.java
@@ -13,7 +13,7 @@ public abstract class Einwohner {
         return this.name;
     }
 
-    public double getEinkommen() {
+    public int getEinkommen() {
         return einkommen;
     }
